Bugfix. Can now accept extreme values for CONT damage and wounds
The previous commit would cause issues formatting values when
either active or reactive caused minimum wounds > max_wounds_shown.
So for example, if a F2F would always cause a minimum of 4 wounds
and the max we are showing is 3, the formatter would not know
how to format it.

Also, Icepool's Again() needs again_end to be set when every
single face of the die has Again. This makes a ton of sense.
again_end is now set to 0, so after 5 repetitions, he will
receive a failure and stop adding.